Add internal APIs for indexing
