+++ api/routes.py
@@ -1,5 +1,6 @@
 from api.modules import (
     attestations,
+    indexing,
     notifications,
     search,
 )
@@ -13,5 +14,6 @@ def add_resources(api, resources, namespace):
 def init_routes(api):
     # add routes for new modules here
     add_resources(api, attestations.resources, '/api/attestations/')
+    add_resources(api, indexing.resources, '/api/indexing/')
     add_resources(api, notifications.resources, '/api/notifications/')
     add_resources(api, search.resources, '/api/search/')
