+++ dev.env
@@ -49,3 +49,5 @@ APNS_APP_BUNDLE_ID=
 
 FCM_API_KEY=
 FCM_TITLE=Origin update
+
+INTERNAL_API_TOKEN=dev-internal-api-token
