+++ logic/search_indexer_service.py
@@ -1,4 +1,9 @@
+import logging
+
+from marshmallow import ValidationError
+
 from logic.search import SearchClient
+from logic.service_utils import SearchIndexingError
 
 
 class SearchIndexer():
@@ -32,3 +37,26 @@ class SearchIndexer():
     def create_or_update_review(self, review_data):
         # TODO(gagan): implement
         pass
+
+    def index_listing(self, listing):
+        """
+        Indexes a listing.
+
+        Args:
+            listing(dict): JSON representation of the listing to index.
+
+        Raises:
+            SearchIndexingError
+        """
+        # Look for the contract's address. It is used as a unique document ID
+        # in the search index.
+        address = listing.get("contract_address")
+        if not address:
+            raise ValidationError("Contract address missing.")
+
+        # TODO(franck): validate listing against schema before indexing it.
+        try:
+            self.client.index_listing(doc_id=address, doc=listing)
+        except Exception as e:
+            logging.error("Indexing failure: %s", str(e))
+            raise SearchIndexingError("Failed indexing listing.")
